Organize deployment modules around configuration and process ownership
