+++ reef/service/assembly.py
@@ -26,7 +26,7 @@ from reef.runtime.base import InferenceRuntime, TrainingRuntime
 from reef.runtime.registry import RuntimeRegistry
 from reef.runtime.settings import TrainingRuntimeSettings
 from reef.service.app import InferenceRetryPolicy, create_app
-from reef.service.deploy.settings import ServiceSettings, service_owned_keys
+from reef.service.deploy.service_config import ServiceSettings, service_owned_keys
 from reef.service.deploy.training import training_deployment_for
 from reef.storage.postgres import PostgresScenarioStorage
 from reef.storage.records import RecordRetention
