--- reef/service/deploy/arguments.py
@@ -1,5 +1,0 @@
-"""Compatibility imports for the shared component configuration parser."""
-
-from reef.core.config import ConfigArgument, ConfigArgumentParser, config_arguments, config_metadata, config_option
-
-__all__ = ["ConfigArgument", "ConfigArgumentParser", "config_arguments", "config_metadata", "config_option"]
