+++ tests/reef_service/test_bridge_readiness.py
@@ -5,7 +5,7 @@ import sys
 
 import pytest
 
-from reef.service.deploy.config import validate_services
+from reef.service.deploy.execution import validate_services
 from reef.service.deploy.orchestrator import _Stack
 from reef.service.deploy.process import ProcessWorker
