+++ tests/reef_service/test_console_cors.py
@@ -10,7 +10,7 @@ from reef.dispatcher import build_default_dispatcher
 from reef.service import assembly
 from reef.service.auth import create_authentication_middleware
 from reef.service.cors import configure_browser_access
-from reef.service.deploy.settings import service_settings_from_config
+from reef.service.deploy.service_config import service_settings_from_config
 from reef.storage.sqlite import SQLiteScenarioStorage
 
 ORIGIN = "https://api.reefinfra.ai"
