+++ tests/reef_service/test_harness_example.py
@@ -21,7 +21,7 @@ from reef.harness.episodes.run import EpisodeResult
 from reef.recipe import load_recipe_config
 from reef.recipe.config import recipe_config_from_mapping
 from reef.recipe.cordis import CordisRecipe
-from reef.service.deploy.settings import service_settings_from_config
+from reef.service.deploy.service_config import service_settings_from_config
 from reef.storage.sqlite import SQLiteRecordStore
 from reef.train.cordis_backend import Mutation
 from reef.train.trainer import Trainer
