+++ tests/reef_service/test_harness_requests_tutorial.py
@@ -19,7 +19,7 @@ from reef_service.config_helpers import load_harness_deployment as load_config
 from reef.dispatcher import training_request_refusal
 from reef.harness.tree.nodes import directive_shaped, secret_shaped
 from reef.recipe.cordis import CordisRecipe
-from reef.service.deploy.settings import service_settings_from_config
+from reef.service.deploy.service_config import service_settings_from_config
 from reef.train.evaluation.evaluators import BackendAlwaysSelectPlugin
 
 REPO_ROOT = Path(__file__).resolve().parents[2]
