+++ tests/reef_service/test_orchestrator_overrides.py
@@ -4,7 +4,8 @@ from __future__ import annotations
 
 import pytest
 
-from reef.service.deploy.orchestrator import InvalidOverrideError, _apply_overrides, _parse_overrides, main
+from reef.service.deploy.cli import InvalidOverrideError, _apply_overrides, _parse_overrides
+from reef.service.deploy.orchestrator import main
 
 
 def test_key_value_pair() -> None:
