+++ tests/reef_service/test_parallel_service_startup.py
@@ -14,7 +14,7 @@ import yaml
 from reef.runtime.executor import Executor
 from reef.runtime.executor.ray import RayExecutor
 from reef.service.deploy import orchestrator
-from reef.service.deploy.config import validate_services
+from reef.service.deploy.execution import validate_services
 from reef.service.deploy.orchestrator import _Stack
 from reef.service.deploy.process import ProcessWorker
